+++ apps/sim/app/(landing)/integrations/data/integrations.json
@@ -13306,7 +13306,35 @@
     "authType": "oauth",
     "category": "tools",
     "integrationTypes": ["communication", "developer-tools"],
-    "tags": ["messaging", "webhooks", "automation"]
+    "tags": ["messaging", "webhooks", "automation"],
+    "landingContent": {
+      "install": {
+        "heading": "Add Sim to your Slack workspace",
+        "intro": "Sim connects to Slack through Slack’s official OAuth flow. The “Add to Slack” button lives inside your Sim account (after sign-in) — connect from there and the Sim bot is installed in your Slack workspace. The steps below show exactly how to reach it.",
+        "steps": [
+          {
+            "title": "Create your free Sim account",
+            "body": "Sign up at sim.ai — no credit card required."
+          },
+          {
+            "title": "Add a Slack block",
+            "body": "Open a workflow, drag in a Slack block, and open its credential dropdown."
+          },
+          {
+            "title": "Connect Slack",
+            "body": "Click Connect Slack, choose your workspace, and approve the requested permissions. This installs the Sim bot in your Slack workspace."
+          },
+          {
+            "title": "Invite the bot and build",
+            "body": "Invite the Sim bot to the channels it should act in, pick a Slack action, wire it into your agent, and run."
+          }
+        ]
+      },
+      "privacy": {
+        "body": "Sim requests only the Slack permissions its actions and triggers need, and never shows private channel names or messages to people who are not members of those channels in Slack.",
+        "href": "/privacy"
+      }
+    }
   },
   {
     "type": "smtp",
